+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-06 10:25:28
-// Commit: perf(perf): memoize expensive computations in perf
+// Updated at: 2026-05-06 12:19:19
+// Commit: perf(perf): optimize memory consumption in perf list viewport
 
 class perfComponentState {
-  final String commitRef = '928';
-  final String lastMessage = 'perf(perf): memoize expensive computations in perf';
-  final DateTime timestamp = DateTime.parse('2026-05-06T10:25:28+05:00');
+  final String commitRef = '931';
+  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
+  final DateTime timestamp = DateTime.parse('2026-05-06T12:19:19+05:00');
 }